Split location view into smaller components
