+++ app/assets/scripts/views/locations-hub.js
@@ -563,7 +563,7 @@ var LocationsHub = createReactClass({
   }
 });
 
-// /////////////////////////////////////////////////////////////////// //
+// /////////////////////////////////////////////////////////////////////
 // Connect functions
 
 function selector (state) {
